// SGUEES-SPA/src/app/pages/Shop/com-tipo-soli-cotiza/models/com-tipo-soli-cotiza.ts
export interface ComTipoSoliCotiza {
	CORR_EMPRESA: number;
	CORR_TIPO_SOLI_COTIZA: number;
	NOMBRE_TIPO_SOLI_COTIZA: string;
	CLASE_SOLI_COTIZA: string;
	NOMBRE_CLASE_SOLI_COTIZA: string;
	USUARIO_CREA: string;
	FECHA_CREA: Date;
	ESTACION_CREA: string;
	USUARIO_ACTU: string;
	FECHA_ACTU: Date;
	ESTACION_ACTU: string;
}